requirements.txt
fastapi
sqlalchemy
psycopg2
pydantic
uvicorn
requests
python-dotenv